Separating dump from import, allowing import to happen remotely

+++ lib/capistrano/mysqldump.rb
@@ -16,14 +16,15 @@ Capistrano::Configuration.instance.load do
       set :mysqldump_location, host || host == "localhost" ? :local : :remote unless exists?(:mysqldump_location)
 
       # for convenience
-      set :mysqldump_filename, "%s-%s.sql" % [application, Time.now.to_i]
+      set :mysqldump_filename, "%s.sql" % application
       set :mysqldump_filename_gz, "%s.gz" % mysqldump_filename
-      set :mysqldump_remote_filename, File.join( mysqldump_remote_tmp_dir, mysqldump_filename_gz )
+      set :mysqldump_remote_filename, File.join( mysqldump_remote_tmp_dir, mysqldump_filename )
+      set :mysqldump_remote_filename_gz, File.join( mysqldump_remote_tmp_dir, mysqldump_filename_gz )
       set :mysqldump_local_filename, File.join( mysqldump_local_tmp_dir, mysqldump_filename )
       set :mysqldump_local_filename_gz, File.join( mysqldump_local_tmp_dir, mysqldump_filename_gz )
     end
 
-    task :dump, :roles => :db do
+    task :dump, :roles => :web do
       setup 
       username, password, database, host = mysqldump_config.values_at *%w( username password database host )
 
@@ -33,14 +34,14 @@ Capistrano::Configuration.instance.load do
       case mysqldump_location
       when :remote
         mysqldump_cmd += " -u %s -p %s" % [ username, database ]
-        mysqldump_cmd += " | gzip > %s" % mysqldump_remote_filename
+        mysqldump_cmd += " | gzip > %s" % mysqldump_remote_filename_gz
 
         run mysqldump_cmd do |ch, stream, out|
           ch.send_data "#{password}\n" if out =~ /^Enter password:/
         end
 
-        download mysqldump_remote_filename, mysqldump_local_filename_gz, :via => :scp
-        run "rm #{mysqldump_remote_filename}"
+        download mysqldump_remote_filename_gz, mysqldump_local_filename_gz, :via => :scp
+        run "rm #{mysqldump_remote_filename_gz}"
         
         `gunzip #{mysqldump_local_filename_gz}`
       when :local
@@ -53,14 +54,30 @@ Capistrano::Configuration.instance.load do
     end
 
     task :import do
-      config = YAML.load_file("config/database.yml")["development"]
-      username, password, database = config.values_at *%w( username password database )
+      setup 
+      config = YAML.load_file("config/database.yml")[rails_env.to_s]
+      username, password, database, host = config.values_at *%w( username password database host )
+      puts "import to host: #{host}"
+      set :mysql_location, host || host == "localhost" ? :local : :remote unless exists?(:mysql_location)
 
       mysql_cmd = "mysql -u#{username}"
       mysql_cmd += " -p#{password}" if password 
-      `#{mysql_cmd} -e "drop database #{database}; create database #{database}"`
-      `#{mysql_cmd} #{database} < #{mysqldump_local_filename}`
-      `rm #{mysqldump_local_filename}`
+
+      case mysql_location
+      when :remote
+        upload mysqldump_local_filename_gz, mysqldump_remote_filename_gz, :via => :scp
+        run "gunzip #{mysqldump_remote_filename_gz}"
+        run "#{mysql_cmd} #{database} < #{mysql_remote_filename}" do |ch, stream, out|
+          ch.send_data "#{password}\n" if out =~ /^Enter password:/
+        end
+        run "rm #{mysqldump_remote_filename}"
+        
+      when :local
+        `#{mysql_cmd} -e "drop database #{database}; create database #{database}"`
+        `#{mysql_cmd} #{database} < #{mysqldump_local_filename}`
+        `rm #{mysqldump_local_filename}`
+      end
+
     end
   end
 end
